pdf-to-md/requirements.txt
# Table detection (vision-based)
pdf2image>=1.16.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
timm>=0.9.0

# Converters
docling>=2.0.0
pymupdf>=1.23.0
pymupdf4llm>=0.0.1
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
